Bump version and other minor changes
- Possibly fixed progress bar going out of bounds.
- Improve book reader controls in landscape mode.
- Book reader no longer displays status bar after every page. Only displayed during zooming or if screen is touched in the middle or outside touch area depending on the orientation.
- Setting archive bit can only be used with dev options enabled.

+++ include/progress_bar.h
@@ -1,6 +1,6 @@
 #ifndef NX_SHELL_MENU_PROGRESS_BAR_H
 #define NX_SHELL_MENU_PROGRESS_BAR_H
 
-void ProgressBar_DisplayProgress(char *msg, char *src, u32 offset, u32 size);
+void ProgressBar_DisplayProgress(char *msg, char *src, u64 offset, u64 size);
 
 #endif
